Rust

+++ src/utils/color.rs
@@ -1,4 +1,3 @@
-#[derive(Default)]
 pub struct Color {
     pub r: f32,
     pub g: f32,
@@ -19,9 +18,9 @@ impl Default for Color {
 
 impl Color {
     pub fn rgba8888to_color (&mut self, value: u32) {
-        self.r = ((value & 0xff000000) >> 24) / 255f32;
-        self.g = ((value & 0x00ff0000) >> 16) / 255f32;
-        self.b = ((value & 0x0000ff00) >> 8) / 255f32;
-        self.a = (value & 0x000000ff) / 255f32;
+        self.r = (((value & 0xff000000) >> 24) / 255) as f32;
+        self.g = (((value & 0x00ff0000) >> 16) / 255) as f32;
+        self.b = (((value & 0x0000ff00) >> 8) / 255) as f32;
+        self.a = ((value & 0x000000ff) / 255) as f32;
     }
 }
